notes - refactoring logic in accordionItem, remove some props, a few changes in useEffect Notes.tsx
