+++ src/utils/getFormattedDate.ts
@@ -1,6 +1,6 @@
 const oneThousand = 1000;
 
-export const getFormatedDate = (dt: string | number) => {
+export const getFormattedDate = (dt: string | number) => {
   const date = new Date(+dt * oneThousand);
   const mm = date.getMonth() + 1;
   const yyyy = date.getFullYear();
