+++ src/utils/index.ts
@@ -1,3 +1,3 @@
 export {getLowerCaseString} from './getLowerCaseString';
 export {getCapitalizedString} from './getCapitalizedString';
-export {getFormatedDate, getDateToTimeStamp} from './getFormatedDate';
+export {getFormattedDate, getDateToTimeStamp} from './getFormattedDate';
