+++ src/widgets/notes/components/accordionWithNotes/accordionItem/typings/index.ts
@@ -2,8 +2,7 @@ import {NotesType} from 'src/widgets/notes/typings';
 
 export type PropsType = {
   city?: string;
-  country: string;
   foundCity: string;
-  notes: NotesType[];
-  removeNote: (title?: string) => void;
+  filteredNotesByCountry: NotesType[];
+  removeNote: (title: string) => void;
 };
