+++ src/widgets/notes/components/notesItem/typings/index.ts
@@ -3,5 +3,5 @@ import {NotesType} from 'src/widgets/notes/typings';
 export type PropsType = {
   city?: string;
   filteredNotesByCountry: NotesType[];
-  removeNote: (title?: string) => void;
+  removeNote: (title: string) => void;
 };
